refactor expression score/part assignment options into a single script with additional menu items
